% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aufenthalt_fun.R
\name{aufenthalt_fun}
\alias{aufenthalt_fun}
\title{aufenthalt_fun}
\usage{
aufenthalt_fun(dt_T, dt_T_1)
}
\arguments{
\item{dt_T}{Target of the shift of hospitalzation status. A data.table featuring individual insurant data according to naming and data type standards used in data set \code{test_data}.}

\item{dt_T_1}{Source of the shift of hospitalization status. A data.table featuring individual insurant data according to naming and data type standards used in data set \code{test_data}.}
}
\value{
A data.table featuring entries of \code{dt_T} and variable \code{Aufenthalt_Vj}
\item{Aufentalt_Vj}{hospitalization status of \code{dt_T_1} assigned to \code{dt_T} by \code{AhvNr}}
}
\description{
Shifts hospitalization status \code{Aufenthalt: J/N} of an \code{AhvNr} from a data set to another. If AhvNr isn't contained in source data, it assigns \code{N} in target. Used in RA calculation to create variable \code{Aufenthalt_Vj} ('previous year's hospitalization status').
}
\examples{
\dontrun{
aufenthalt_fun(
test_data[Jahr == 2019 & MonateHorizont == 14],
test_data[Jahr == 2018 & MonateHorizont == 26])
}
}
\seealso{
\url{https://www.kvg.org/api/rm/S443P62E88YU756/200420-kuv-va-berechnungsformeln-fuer-den-risikoau.pdf}
}
\author{
Magnus Vieten, Yannick Schwarz
}
