% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_ra.R
\name{compute_ra}
\alias{compute_ra}
\title{Run Risk Adjustment calculations}
\usage{
compute_ra(
  Year,
  ind_data = NULL,
  pcg_list = NULL,
  Clean_Steps = FALSE,
  export_objects = "all"
)
}
\arguments{
\item{Year}{An integer specifying the year of the risk adjustment}

\item{ind_data}{\code{data.frame} or \code{data.table} object. Uses synthetic
test data included in the package if NULL.}

\item{pcg_list}{Either a path to an .xlsx with the PCG list, or a \code{data.frame}/\code{data.table}
Uses the public PCG list of year 2022 if NULL.}

\item{Clean_Steps}{Boolean. Whether to clean up intermediate results.}

\item{export_objects}{A character vector naming objects to export. By default, all objects are exported}
}
\value{
A \code{list} of objects specified in export_objects.
}
\description{
Runs all steps of the risk adjustment calculations and returns the results
as a named list.
}
\examples{
\dontrun{
res <- compute_ra(2022, Clean_Steps = TRUE)
}
}
\author{
Magnus Vieten, Yannick Schwarz, Thomas Fischer
}
