% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\name{globals}
\alias{globals}
\title{Define Global Variables}
\usage{
globals()
}
\description{
Round up all unbound variables, such that the note 'no visible binding for
global variable' when checking the package disappears.
}
\seealso{
\href{https://community.rstudio.com/t/how-to-solve-no-visible-binding-for-global-variable-note/28887}{Link1}, \href{https://www.r-bloggers.com/2019/08/no-visible-binding-for-global-variable/}{link2} and \href{https://stackoverflow.com/questions/40251801/how-to-use-utilsglobalvariables}{link3} # nolint: line_length_linter.
}
