% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_age_groups.R
\name{make_age_groups}
\alias{make_age_groups}
\title{Make age groups}
\usage{
make_age_groups(DT, style = "normal")
}
\arguments{
\item{DT}{A \code{data.table} containing variables "Jahr" and "Geburtsjahr" (both of type integer).}

\item{style}{Designate output format. \code{"normal"} gives type "36", \code{"sora"} gives type "36-40".}
}
\value{
Adds variable "Altersklasse" to input \code{data.table}
}
\description{
Make age groups in accordance to Art. 2 VORA from 19.10.2016 (Release 1.1.2022).
}
\examples{
make_age_groups(data.table::data.table(Jahr = 2021, Geburtsjahr = 1962))

}
\seealso{
\href{https://www.fedlex.admin.ch/eli/cc/2016/674/de}{VORA}
}
\author{
Yannick Schwarz
}
