% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcg_grouping.R
\name{pcg_grouping}
\alias{pcg_grouping}
\title{PCG Grouping}
\usage{
pcg_grouping(data_input, pcg_liste, inter_returns = NULL)
}
\arguments{
\item{data_input}{A \code{data.table} or \code{data.frame} featuring RA
individual data according to naming and data type standards used in PB. For
renaming use function \code{GEPack::sora_naming}.}

\item{pcg_liste}{\code{data.table} or \code{data.frame} with the PCG-list or
character vector with the path to an excel file.}

\item{inter_returns}{Character vector with return objects from intermediate
steps of the function. Possible values are "data_input", "ddd_groupings",
"input_fun", "last_step", "pcg_gtin", "pcg_li", "pcg_li_all", "pcg_liste",
"pcg_pharm", "pcg_threshold", "show_all_steps".}
}
\value{
A \code{data.table} containing
\item{Jahr}{Data year}
\item{MonateHorizont}{Horizon of data status, either \code{14}, \code{26}
or \code{NA}}
\item{AhvNr}{Pseudo SSN}
\item{Geburtsjahr}{Year of birthday - used with \code{AhvNr} to warrant
identification}
\item{pcg}{Grouped PCG}
}
\description{
Groups individual data into PCGs
}
\examples{
\dontrun{
pcg_grouping(
  data_input = some_RA_IndData_object,
  pcg_liste = "path/to/prepped/pcg_list.xlsx"
)
}
}
\author{
Magnus Vieten, Yannick Schwarz
}
