% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcg_hierarchy.R
\name{pcg_hierarchy}
\alias{pcg_hierarchy}
\title{PCG Hierarchisation}
\usage{
pcg_hierarchy(data_input, direction, display_all = FALSE, display_hyp = FALSE)
}
\arguments{
\item{data_input}{A \code{data.table} of grouped PCG data with necessary
variables \code{AhvNr}, \code{Geburtsjahr}, \code{pcg}.}

\item{direction}{Character vector which determines the output dimensions.
\code{"wide"} returns a dummy matrix output, \code{"long"} returns a long
format output.}

\item{display_all}{Boolean, whether to return all variables, even if no
corresponding PCG entries exist. Default: FALSE.}

\item{display_hyp}{Logical with default \code{FALSE}. If \code{TRUE}
non-independent PCG "hyp" is treated as independent (and also still combined
with DM2 to DM2hyp).}
}
\value{
A \code{data.table} containing
\item{AhvNr}{Pseudo SSN}
\item{Geburtsjahr}{Year of birthday - used with \code{AhvNr}
to warrant identification}
\item{pcg}{Grouped PCG}
}
\description{
Apply hierarchy to grouped PCG data. Application accords to
\href{https://www.fedlex.admin.ch/eli/cc/2019/618/de#art_2}{Art. 2 & 3 VORA-EDI}
}
\examples{
\dontrun{
pcg_hierarchy(
  data_input = grouped_pcg_dt,
  pcg_liste = "path/to/prepped/pcg_list.xlsx",
  direction = "wide",
  display_all = TRUE,
  display_hyp = TRUE
)
}
}
\seealso{
\href{https://www.fedlex.admin.ch/eli/cc/2019/618/de#art_2}{Art. 2 & 3 VORA-EDI}
}
\author{
Magnus Vieten, Yannick Schwarz
}
