% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rg_grid.R
\name{rg_grid}
\alias{rg_grid}
\title{create RA risk group grid}
\usage{
rg_grid(children = FALSE, kantonal = TRUE)
}
\arguments{
\item{children}{Logical with default \code{FALSE}. If \code{TRUE} age group \code{"0-18"} is added to the grid.}

\item{kantonal}{Logical with default \code{TRUE} where Swiss cantons are added to the grid. \code{FALSE} returns risk group grid without cantons.}
}
\value{
A \code{data.table} containing all combinations of
\item{Kanton}{26 swiss cantons}
\item{Altersklasse}{15 age groups (16 when \code{children = T})}
\item{Geschlecht}{M/F}
\item{Aufenthalt}{J/N dummy variable indicating whether person was hospitalized for at least 3 consecutive nights}
}
\description{
Running \code{rg_grid()} returns a \code{data.table} of all 1560 risk groups featured in current RA.
}
\examples{
rg_grid()
}
\seealso{
\href{https://www.fedlex.admin.ch/eli/cc/2016/674/de#art_1}{Art. 1 & 2 VORA}
}
\author{
Magnus Vieten, Yannick Schwarz
}
