% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sora_datatype.R
\name{sora_datatype}
\alias{sora_datatype}
\title{Make the same datatypes as in SORA PCG}
\usage{
sora_datatype(data, Ahv_Char = TRUE)
}
\arguments{
\item{data}{A \code{data.table} object.}

\item{Ahv_Char}{Logical if \code{TRUE} AhvNr should be a character vector or if \code{FALSE} a numeric vector.}
}
\description{
Make the same datatypes as in SORA PCG
}
