% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teuerung_fun.R
\name{teuerung_fun}
\alias{teuerung_fun}
\title{Teuerung}
\usage{
teuerung_fun(dt_T_14, dt_T_1_14)
}
\arguments{
\item{dt_T_14}{A \code{data.table} featuring individual insurance data according to naming and data type standards used in data set \code{test_data}.}

\item{dt_T_1_14}{A \code{data.table} featuring individual insurance data according to naming and data type standards used in data set \code{test_data}.}
}
\value{
A \code{data.table} containing
\item{Kanton}{Swiss canton}
\item{pi_Niv_k}{Total cost increase component divided by structural cost increase component ('Niveateuerung')}
\item{pi_Str_k1}{Structural cost increase component}
\item{pi_Ges_k1}{Total cost increase component}
}
\description{
Calculates the cost increase ('Niveauteuerung pro Kanton') used as an extrapolation factor in risk equalization. Derivation see: \href{https://www.kvg.org/de/risikoausgleich-pcg-_content---1--3125.html}{Berechnungsformeln fuer den Risikoasugleich mit PCG ab 2020, Section B}.
}
\examples{
\dontrun{
pcg_grouping(data_input = some_RA_IndData_object, pcg_liste = "path/to/prepped/pcg_list.xlsx")
}
}
\seealso{
\url{https://www.kvg.org/api/rm/S443P62E88YU756/200420-kuv-va-berechnungsformeln-fuer-den-risikoau.pdf}
}
\author{
Magnus Vieten, Yannick Schwarz
}
