% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vw_fun.R
\name{vw_fun}
\alias{vw_fun}
\title{Versichererwechsler}
\usage{
vw_fun(T_1_26, T_14)
}
\arguments{
\item{T_1_26}{A \code{data.table} featuring RA individual data according to naming and data type standards used in RA. 1st component of comparison ('previous year')}

\item{T_14}{A \code{data.table} featuring RA individual data according to naming and data type standards used in RA. 2nd component of comparison.}
}
\value{
A \code{data.table} containing
\item{BagNr}{Integer with the BAG-Nr.}
\item{AhvNr}{Pseudo SSN}
\item{Geburtsjahr}{Year of birthday - used with \code{AhvNr} to warrant identification}
\item{VW}{J/N dummy variable indicating the status of insurance change}
}
\description{
Returns status of insurance change \code{VW} for an \code{AhvNr} between two RA data sets. See determination rules in \href{https://www.kvg.org/de/datenerhebung-_content---1--1053.html}{Benutzeranleitung SORA PCG, Kap. 10}.
}
\examples{
\dontrun{
vw_fun(T_1_26_auf, T_14_auf)
}
}
\author{
Magnus Vieten, Yannick Schwarz
}
